Extract JIT device config enumeration into jit_build
Move `enumerate_jit_device_configs` and its `dispatch_message_addr`
helper from `tt_metal/tools/precompile_fw/precompile_fw.cpp` into
`tt_metal/jit_build/jit_device_config.{hpp,cpp}` so the enumeration
becomes reusable by the forthcoming offline kernel compilation API,
not just firmware precompilation.

The enumerator gains a `soc_descriptor_path` parameter and derives the
base DRAM bank count from the `dram_views` list in that YAML, replacing
the arch->literal `num_dram_banks_map`. Values are unchanged today
(WH=12, BH=8) but are now sourced from the same file the runtime
`metal_SocDescriptor` path reads.

Hand-curated per-product routing and DRAM-harvesting coverage is
isolated into two named helpers (`routing_fw_configs_for_product`,
`dram_harvesting_configs_for_product`) with a tracking TODO. The
residual `pcie_core_map` becomes `pcie_core_for_arch`. A defensive
`TT_FATAL` guards `dram_harvesting_count <= base_num_dram_banks`.

`copy_firmware_to_precompiled_dir`, `precompile_for_config`, and
`main()` stay in `precompile_fw.cpp`. Move `yaml-cpp::yaml-cpp` from
`precompile_fw`'s `target_link_libraries` to `jit_build`. Drop the
obsolete "TODO: Add a factory method to create JitDeviceConfig from a
YAML profile" note; `enumerate_jit_device_configs` is that factory.

Made-with: Cursor
