+++ tt_metal/tools/precompile_fw/precompile_fw.cpp
@@ -6,33 +6,14 @@
 #include <filesystem>
 #include <string>
 
-#include <yaml-cpp/yaml.h>
+#include "jit_build/build.hpp"
 #include "jit_build/build_env_manager.hpp"
 #include "jit_build/jit_device_config.hpp"
-#include "jit_build/build.hpp"
-#include "llrt/hal.hpp"
 #include "llrt/rtoptions.hpp"
 
 namespace tt::tt_metal {
 
 namespace {
-
-// FIXME: use DispatchMemMap::get_dispatch_message_addr_start() once it is decoupled from global MetalContext
-uint32_t dispatch_message_addr(const Hal& hal, DispatchCoreType dispatch_core_type) {
-    uint32_t stream_index = 0;
-    if (dispatch_core_type == DispatchCoreType::WORKER) {
-        // There are 64 streams. CBs use entries 8-39.
-        stream_index = 48u;
-    } else if (dispatch_core_type == DispatchCoreType::ETH) {
-        // There are 32 streams.
-        stream_index = 16u;
-    } else {
-        TT_THROW("dispatch_message_addr not implemented for core type");
-    }
-
-    return hal.get_noc_overlay_start_addr() + (hal.get_noc_stream_reg_space_size() * stream_index) +
-           (hal.get_noc_stream_remote_dest_buf_space_available_update_reg_index() * sizeof(uint32_t));
-}
 
 void copy_firmware_to_precompiled_dir(
     const std::string& firmware_out_path, const std::string& precompiled_firmware_dir) {
@@ -48,117 +29,6 @@ void copy_firmware_to_precompiled_dir(
 }
 
 }  // namespace
-
-void enumerate_jit_device_configs(
-    tt::ARCH arch, const std::string& core_descriptor_path, const std::function<void(const JitDeviceConfig&)>& callback) {
-    // FIXME: hardcoded values
-    constexpr uint32_t profiler_dram_bank_size_per_risc_bytes = 0;
-    // Only support compiling 2-ERISC mode for Blackhole.
-    const bool enable_2_erisc_mode = (arch == tt::ARCH::BLACKHOLE);
-
-    const bool enable_dram_backed_cq = false;
-
-    // FIXME: hardcoded based on UMD values
-    static const std::unordered_map<tt::ARCH, uint32_t> num_dram_banks_map = {
-        {tt::ARCH::WORMHOLE_B0, 12},
-        {tt::ARCH::BLACKHOLE, 8},
-    };
-    static const std::unordered_map<tt::ARCH, CoreCoord> pcie_core_map = {
-        {tt::ARCH::WORMHOLE_B0, {0, 3}},
-        {tt::ARCH::BLACKHOLE, {19, 24}},
-    };
-    const CoreCoord pcie_core = pcie_core_map.at(arch);
-
-    YAML::Node core_descriptor_yaml = YAML::LoadFile(core_descriptor_path);
-    for (const auto& product : core_descriptor_yaml) {
-        const std::string product_name = product.first.as<std::string>();
-        std::vector<bool> routing_fw_enabled_cfgs = {false};
-        std::vector<uint32_t> dram_harvesting_cfgs = {0};
-        // FIXME: hardcoded logic adopted from core_descriptor.cpp
-        if (arch == tt::ARCH::WORMHOLE_B0) {
-            routing_fw_enabled_cfgs = {true};
-            if (product_name == "galaxy" || product_name == "nebula_x1") {
-                routing_fw_enabled_cfgs.push_back(false);
-            }
-        }
-        // FIXME: hardcoded values adopted from expected_dram_harvested_units_map in UMD
-        if (arch == tt::ARCH::BLACKHOLE && product_name == "2xharvested") {
-            dram_harvesting_cfgs.push_back(1);
-        }
-        for (const auto& axis_config : product.second) {
-            std::string dispatch_core_axis_str = axis_config.first.as<std::string>();
-            tt_metal::DispatchCoreAxis dispatch_core_axis;
-            if (dispatch_core_axis_str == "row") {
-                dispatch_core_axis = tt_metal::DispatchCoreAxis::ROW;
-            } else if (dispatch_core_axis_str == "col") {
-                dispatch_core_axis = tt_metal::DispatchCoreAxis::COL;
-            } else {
-                TT_THROW("Invalid dispatch core axis: {}", dispatch_core_axis_str);
-            }
-            for (const auto& config_node : axis_config.second) {
-                auto num_hw_cqs = config_node.first.as<uint8_t>();
-                const auto& config = config_node.second;
-
-                // FIXME: hardcoded logic adopted from core_descriptor.cpp
-                // tg_compute_with_storage_grid_range is not to be supported.
-                auto compute_with_storage_start = config["compute_with_storage_grid_range"]["start"];
-                auto compute_with_storage_end = config["compute_with_storage_grid_range"]["end"];
-                size_t end_x = compute_with_storage_end[0].as<size_t>();
-                size_t end_y = compute_with_storage_end[1].as<size_t>();
-                size_t start_x = compute_with_storage_start[0].as<size_t>();
-                size_t start_y = compute_with_storage_start[1].as<size_t>();
-                auto compute_grid_size = CoreCoord((end_x - start_x) + 1, (end_y - start_y) + 1);
-                auto num_l1_banks = compute_grid_size.x * compute_grid_size.y;
-
-                auto dispatch_core_type_str = config["dispatch_core_type"].as<std::string>();
-                tt_metal::DispatchCoreType dispatch_core_type;
-                if (dispatch_core_type_str == "tensix") {
-                    dispatch_core_type = tt_metal::DispatchCoreType::WORKER;
-                } else if (dispatch_core_type_str == "ethernet") {
-                    dispatch_core_type = tt_metal::DispatchCoreType::ETH;
-                } else {
-                    TT_THROW("Invalid dispatch core type: {}", dispatch_core_type_str);
-                }
-
-                // now enumerate all possible combinations of routing_fw_enabled and dram_harvesting_cfg
-                for (const auto& routing_fw_enabled : routing_fw_enabled_cfgs) {
-                    for (const auto& dram_harvesting_count : dram_harvesting_cfgs) {
-                        const uint32_t num_dram_banks = num_dram_banks_map.at(arch) - dram_harvesting_count;
-                        tt::tt_metal::Hal hal(
-                            arch,
-                            routing_fw_enabled,
-                            enable_2_erisc_mode,
-                            profiler_dram_bank_size_per_risc_bytes,
-                            enable_dram_backed_cq,
-                            /*is_simulator=*/false,
-                            /*enable_blackhole_dram_programmable_cores=*/true);
-                        JitDeviceConfig jit_device_config = {
-                            .hal = &hal,
-                            .arch = arch,
-                            .num_dram_banks = num_dram_banks,
-                            .num_l1_banks = num_l1_banks,
-                            .pcie_core = pcie_core,
-                            // We only precompile with coordinate_virtualization_enabled = true, so harvesting_mask
-                            // has no effect on compilation
-                            .harvesting_mask = 0,
-                            .dispatch_core_type = dispatch_core_type,
-                            .dispatch_core_axis = dispatch_core_axis,
-                            .coordinate_virtualization_enabled = true,
-                            .dispatch_message_addr = dispatch_message_addr(hal, dispatch_core_type),
-                            .max_cbs = hal.get_arch_num_circular_buffers(),
-                            .num_hw_cqs = num_hw_cqs,
-                            .routing_fw_enabled = routing_fw_enabled,
-                            // We only precompile with profiler disabled, so profiler_dram_bank_size_per_risc_bytes
-                            // has no effect on compilation
-                            .profiler_dram_bank_size_per_risc_bytes = profiler_dram_bank_size_per_risc_bytes,
-                        };
-                        callback(jit_device_config);
-                    }
-                }
-            }
-        }
-    }
-}
 
 void precompile_for_config(
     const tt::tt_metal::JitDeviceConfig& jit_device_config, const tt::llrt::RunTimeOptions& rtoptions) {
@@ -182,15 +52,16 @@ namespace {
 struct PrecompileConfig {
     tt::ARCH arch;
     std::string core_descriptor_name;
+    std::string soc_descriptor_name;
 };
 
 const auto supported_configs = std::to_array<PrecompileConfig>({
-    {tt::ARCH::WORMHOLE_B0, "wormhole_b0_80_arch.yaml"},
-    {tt::ARCH::WORMHOLE_B0, "wormhole_b0_80_arch_eth_dispatch.yaml"},
-    {tt::ARCH::WORMHOLE_B0, "wormhole_b0_80_arch_fabric_mux.yaml"},
-    {tt::ARCH::BLACKHOLE, "blackhole_140_arch.yaml"},
-    {tt::ARCH::BLACKHOLE, "blackhole_140_arch_eth_dispatch.yaml"},
-    {tt::ARCH::BLACKHOLE, "blackhole_140_arch_fabric_mux.yaml"},
+    {tt::ARCH::WORMHOLE_B0, "wormhole_b0_80_arch.yaml", "wormhole_b0_80_arch.yaml"},
+    {tt::ARCH::WORMHOLE_B0, "wormhole_b0_80_arch_eth_dispatch.yaml", "wormhole_b0_80_arch.yaml"},
+    {tt::ARCH::WORMHOLE_B0, "wormhole_b0_80_arch_fabric_mux.yaml", "wormhole_b0_80_arch.yaml"},
+    {tt::ARCH::BLACKHOLE, "blackhole_140_arch.yaml", "blackhole_140_arch.yaml"},
+    {tt::ARCH::BLACKHOLE, "blackhole_140_arch_eth_dispatch.yaml", "blackhole_140_arch.yaml"},
+    {tt::ARCH::BLACKHOLE, "blackhole_140_arch_fabric_mux.yaml", "blackhole_140_arch.yaml"},
 });
 
 }  // namespace
@@ -207,10 +78,15 @@ int main() {
     // cache so the .elf and its weakened ELF stay consistent before the copy.
     rtoptions.set_disable_precompiled_fw(true);
     const std::string core_descriptors_dir = rtoptions.get_root_dir() + "tt_metal/core_descriptors/";
-    for (const auto& [arch, core_descriptor_name] : supported_configs) {
+    const std::string soc_descriptors_dir = rtoptions.get_root_dir() + "tt_metal/soc_descriptors/";
+    for (const auto& [arch, core_descriptor_name, soc_descriptor_name] : supported_configs) {
         std::string full_core_descriptor_path = core_descriptors_dir + core_descriptor_name;
+        std::string full_soc_descriptor_path = soc_descriptors_dir + soc_descriptor_name;
         tt::tt_metal::enumerate_jit_device_configs(
-            arch, full_core_descriptor_path, [&rtoptions](const tt::tt_metal::JitDeviceConfig& jit_device_config) {
+            arch,
+            full_core_descriptor_path,
+            full_soc_descriptor_path,
+            [&rtoptions](const tt::tt_metal::JitDeviceConfig& jit_device_config) {
                 tt::tt_metal::precompile_for_config(jit_device_config, rtoptions);
             });
     }
